% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting.R
\name{asymmetric_laplace}
\alias{asymmetric_laplace}
\title{Asymmetric Laplace for Quantile Regression}
\usage{
asymmetric_laplace()
}
\value{
An object of class \code{"family"} representing the Asymmetric Laplace distribution.
}
\description{
Asymmetric Laplace for Quantile Regression
}
