% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional-families.R
\name{beta_binomial}
\alias{beta_binomial}
\title{Beta Binomial Family for Overdispersed Binary Data}
\usage{
beta_binomial(link = "logit")
}
\arguments{
\item{link}{Link function for probability parameter (default: "logit")}
}
\value{
A family object of class \code{"family"}.
}
\description{
Beta Binomial Family for Overdispersed Binary Data
}
