% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional-families.R
\name{beta_variants}
\alias{beta_variants}
\alias{beta0}
\alias{beta1}
\alias{logitbeta}
\title{Alternative Beta Parameterizations}
\usage{
beta0(link = "logit")

beta1(link = "logit")

logitbeta(link = "logit")
}
\arguments{
\item{link}{Link function (default: "logit")}
}
\description{
Alternative parameterizations of the beta distribution for specialized use cases.
}
\details{
\itemize{
\item beta0: Beta distribution parameterized for boundary behavior
\item beta1: Alternative beta parameterization for different tail behavior
\item logitbeta: Beta with explicit logit parameterization
}
}
