% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_tables.R
\name{build_html_table_styled}
\alias{build_html_table_styled}
\title{Build HTML Table with Enhanced Styling}
\usage{
build_html_table_styled(
  model_info,
  all_predictors,
  display_predictors,
  dv.labels,
  show.ci,
  show.rope,
  show.p_sig,
  show.pd,
  show.bf,
  show.intercept,
  show.r2,
  show.icc,
  show.nobs,
  digits,
  title,
  bootstrap,
  table.style,
  font.family,
  font.size,
  header.bg,
  stripe.bg,
  CSS,
  verbose = FALSE
)
}
\description{
Build HTML Table with Enhanced Styling
}
\keyword{internal}
