% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{cauchy}
\alias{cauchy}
\title{Specify Cauchy Prior Distribution}
\usage{
cauchy(location = 0, scale = 1)
}
\arguments{
\item{location}{Location parameter (default 0)}

\item{scale}{Scale parameter (default 1)}
}
\value{
A prior distribution object
}
\description{
Specify Cauchy Prior Distribution
}
