% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_convergence}
\alias{check_convergence}
\title{Quick model diagnostics}
\usage{
check_convergence(object)
}
\arguments{
\item{object}{A qbrms_fit object.}
}
\value{
Invisible TRUE if successful.
}
\description{
Quick model diagnostics
}
