% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional-families.R
\name{circular_normal}
\alias{circular_normal}
\alias{von_mises}
\title{Circular Normal Family for Directional Data}
\usage{
circular_normal(link = "tan_half", link.kappa = "log")

von_mises(link = "tan_half", link.kappa = "log")
}
\arguments{
\item{link}{Link function for mean direction}

\item{link.kappa}{Link function for concentration}
}
\value{
A family object of class \code{"family"}.
}
\description{
Circular Normal Family for Directional Data
}
