% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_methods.R
\name{clean_coefficient_names}
\alias{clean_coefficient_names}
\title{Clean up malformed coefficient names from mixed effects models}
\usage{
clean_coefficient_names(coef_table, verbose = FALSE)
}
\description{
Clean up malformed coefficient names from mixed effects models
}
\keyword{internal}
