% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_methods.R
\name{coef.qbrms_fit}
\alias{coef.qbrms_fit}
\title{Extract Coefficients from qbrms Models}
\usage{
\method{coef}{qbrms_fit}(object, ...)
}
\arguments{
\item{object}{A qbrms_fit object}

\item{...}{Additional arguments (unused)}
}
\value{
Named vector of coefficients
}
\description{
Extract Coefficients from qbrms Models
}
