% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinal_tmb.R
\name{coef.tmb_ordinal_qbrms_fit}
\alias{coef.tmb_ordinal_qbrms_fit}
\title{Coefficients Method for TMB Ordinal Fits}
\usage{
\method{coef}{tmb_ordinal_qbrms_fit}(object, ...)
}
\arguments{
\item{object}{A tmb_ordinal_qbrms_fit object}

\item{...}{Additional arguments}
}
\value{
Named vector of coefficients
}
\description{
Coefficients Method for TMB Ordinal Fits
}
