% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_effects.R
\name{conditional_effects}
\alias{conditional_effects}
\title{Conditional Effects for qbrms Models}
\usage{
conditional_effects(object, ...)
}
\arguments{
\item{object}{Model object}

\item{...}{Additional arguments}
}
\value{
Object of class "brms_conditional_effects"
}
\description{
Generic function for computing conditional effects
}
