% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family_conversion.R
\name{convert_binomial_formula}
\alias{convert_binomial_formula}
\title{Convert Binomial Formulas Using 'trials()' Syntax}
\usage{
convert_binomial_formula(formula, data, formula_info, verbose = TRUE)
}
\description{
Convert Binomial Formulas Using 'trials()' Syntax
}
\keyword{internal}
