% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinal_tmb.R
\name{create_ordinal_qbrms_result}
\alias{create_ordinal_qbrms_result}
\title{Create qbrms-Compatible Result Object}
\usage{
create_ordinal_qbrms_result(
  tmb_fit,
  formula,
  data,
  family,
  prior,
  verbose = FALSE
)
}
\arguments{
\item{tmb_fit}{TMB fit results}

\item{formula}{Original formula}

\item{data}{Original data}

\item{family}{Model family}

\item{prior}{Prior specifications}

\item{verbose}{Logical; print progress messages}
}
\value{
qbrms-compatible result object
}
\description{
Create qbrms-Compatible Result Object
}
\keyword{internal}
