% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting_helpers.R
\name{create_quantile_fit}
\alias{create_quantile_fit}
\title{Utility operator}
\usage{
create_quantile_fit(formula, data, quantile = 0.5, verbose = TRUE)
}
\description{
Utility operator
}
\keyword{Create}
\keyword{a}
\keyword{fit}
\keyword{for}
\keyword{internal}
\keyword{quantile}
\keyword{regression}
\keyword{simple}
