% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-test-utilities.R
\name{create_test_fit}
\alias{create_test_fit}
\title{Create test fit object for use in tests}
\usage{
create_test_fit()
}
\description{
Create test fit object for use in tests
}
\keyword{internal}
