% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting.R
\name{cumulative}
\alias{cumulative}
\title{Cumulative Family for Ordinal Regression}
\usage{
cumulative(link = "logit")
}
\arguments{
\item{link}{Link function (default: "logit")}
}
\description{
Cumulative Family for Ordinal Regression
}
