% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_visualisation.R
\name{.add_reference_distributions}
\alias{.add_reference_distributions}
\title{Add Reference Distributions}
\usage{
.add_reference_distributions(p, xlim, prior_info)
}
\description{
Add Reference Distributions
}
\keyword{internal}
