% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed_binomial_enhanced.R
\name{.analyse_group_structure}
\alias{.analyse_group_structure}
\title{Analyse grouping structure}
\usage{
.analyse_group_structure(formula, data, y, min_group_size, verbose)
}
\description{
Analyse grouping structure
}
\keyword{internal}
