% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbrm.R
\name{.augment_data_for_stability}
\alias{.augment_data_for_stability}
\title{Data augmentation for model stability}
\usage{
.augment_data_for_stability(data, formula, verbose = TRUE)
}
\description{
Data augmentation for model stability
}
\keyword{internal}
