% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed_binomial_enhanced.R
\name{.augment_data_intelligent}
\alias{.augment_data_intelligent}
\title{Intelligent data augmentation}
\usage{
.augment_data_intelligent(formula, data, diagnostics, verbose)
}
\description{
Intelligent data augmentation
}
\keyword{internal}
