% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesR2.R
\name{.bayes_R2_brms_style}
\alias{.bayes_R2_brms_style}
\title{Compute Bayesian R-squared exactly as in brms}
\usage{
.bayes_R2_brms_style(y, ypred)
}
\description{
Compute Bayesian R-squared exactly as in brms
}
\keyword{internal}
