% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_diagnostics.R
\name{.check_influential_observations}
\alias{.check_influential_observations}
\title{Check for Influential Observations}
\usage{
.check_influential_observations(model)
}
\description{
Check for Influential Observations
}
\keyword{internal}
