% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed_binomial_enhanced.R
\name{.choose_fitting_strategy}
\alias{.choose_fitting_strategy}
\title{Choose optimal fitting strategy}
\usage{
.choose_fitting_strategy(diagnostics, verbose)
}
\description{
Choose optimal fitting strategy
}
\keyword{internal}
