% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinal_tmb.R
\name{.compute_robust_vcov}
\alias{.compute_robust_vcov}
\title{Compute Robust Variance-Covariance Matrix - CORRECTED VERSION}
\usage{
.compute_robust_vcov(obj, opt, verbose = FALSE)
}
\arguments{
\item{obj}{TMB object}

\item{opt}{Optimization result}

\item{verbose}{Logical; print diagnostics}
}
\value{
Variance-covariance matrix with method attribute
}
\description{
Compute Robust Variance-Covariance Matrix - CORRECTED VERSION
}
\keyword{internal}
