% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed_binomial_enhanced.R
\name{.convert_to_inla_formula}
\alias{.convert_to_inla_formula}
\title{Convert lme4-style formula to INLA format}
\usage{
.convert_to_inla_formula(formula)
}
\description{
Convert lme4-style formula to INLA format
}
\keyword{internal}
