% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed_binomial_enhanced.R
\name{.create_balanced_augmentation}
\alias{.create_balanced_augmentation}
\title{Create balanced augmented observations}
\usage{
.create_balanced_augmentation(data, formula, n_aug, diagnostics, verbose)
}
\description{
Create balanced augmented observations
}
\keyword{internal}
