% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_checks.R
\name{.create_density_base_plot}
\alias{.create_density_base_plot}
\title{Create Base Density Plot}
\usage{
.create_density_base_plot(
  plot_densities,
  colours,
  alpha_levels,
  parameter,
  verbose
)
}
\description{
Create Base Density Plot
}
\keyword{internal}
