% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/families.R
\name{.create_ordinal_routing_object}
\alias{.create_ordinal_routing_object}
\title{Create Ordinal Routing Object}
\usage{
.create_ordinal_routing_object(family, family_name)
}
\arguments{
\item{family}{Original family object}

\item{family_name}{Extracted family name}
}
\value{
Routing object for ordinal regression
}
\description{
Create Ordinal Routing Object
}
\keyword{internal}
