% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_checks.R
\name{.create_prior_only_plot}
\alias{.create_prior_only_plot}
\title{Create Prior-Only Plot}
\usage{
.create_prior_only_plot(
  yrep,
  type = "dens_overlay",
  ndraws = NULL,
  verbose = TRUE
)
}
\description{
Create Prior-Only Plot
}
\keyword{internal}
