% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_checks.R
\name{.create_synthetic_data}
\alias{.create_synthetic_data}
\title{Create Synthetic Data for Prior Checks}
\usage{
.create_synthetic_data(formula, n_obs, predictor_values = NULL, verbose = TRUE)
}
\description{
Create Synthetic Data for Prior Checks
}
\keyword{internal}
