% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbrm.R
\name{.diagnose_binomial_issues}
\alias{.diagnose_binomial_issues}
\title{Internal function to diagnose binomial issues}
\usage{
.diagnose_binomial_issues(formula, data, verbose = TRUE)
}
\description{
Internal function to diagnose binomial issues
}
\keyword{internal}
