% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting_helpers.R
\name{.drop_random_effects}
\alias{.drop_random_effects}
\title{Drop random-effect terms from a formula}
\usage{
.drop_random_effects(fml)
}
\description{
Drop random-effect terms from a formula
}
\keyword{internal}
