% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesR2.R
\name{.drop_random_effects_for_r2}
\alias{.drop_random_effects_for_r2}
\title{Remove random effects from formula for design matrix creation}
\usage{
.drop_random_effects_for_r2(formula)
}
\description{
Remove random effects from formula for design matrix creation
}
\keyword{internal}
