% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesR2.R
\name{.estimate_random_effects_variance_from_data_corrected}
\alias{.estimate_random_effects_variance_from_data_corrected}
\title{Estimate random effects variance from data - CORRECTED}
\usage{
.estimate_random_effects_variance_from_data_corrected(object, verbose)
}
\description{
Estimate random effects variance from data - CORRECTED
}
\keyword{internal}
