% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_visualisation.R
\name{.evaluate_prior_density}
\alias{.evaluate_prior_density}
\title{Evaluate Prior Density}
\usage{
.evaluate_prior_density(x, prior_info)
}
\description{
Evaluate Prior Density
}
\keyword{internal}
