% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_export.R
\name{.export_as_markdown}
\alias{.export_as_markdown}
\title{Export as Markdown}
\usage{
.export_as_markdown(model, include_data, include_diagnostics, is_fitted)
}
\description{
Export as Markdown
}
\keyword{internal}
