% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_export.R
\name{.export_as_r_script}
\alias{.export_as_r_script}
\title{Export as R Script}
\usage{
.export_as_r_script(model, include_data, include_diagnostics, is_fitted)
}
\description{
Export as R Script
}
\keyword{internal}
