% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_export.R
\name{.export_as_text}
\alias{.export_as_text}
\title{Export as Plain Text}
\usage{
.export_as_text(model, include_data, include_diagnostics, is_fitted)
}
\description{
Export as Plain Text
}
\keyword{internal}
