% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesR2.R
\name{.extract_inla_fitted_with_random_effects}
\alias{.extract_inla_fitted_with_random_effects}
\title{Extract INLA fitted values that include random effects - CORRECTED}
\usage{
.extract_inla_fitted_with_random_effects(object, ndraws, verbose)
}
\description{
Extract INLA fitted values that include random effects - CORRECTED
}
\keyword{internal}
