% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_checks.R
\name{.extract_parameter_densities}
\alias{.extract_parameter_densities}
\title{Extract Parameter Distribution Densities}
\usage{
.extract_parameter_densities(
  object,
  parameter,
  show_prior,
  ndraws,
  prior_ndraws,
  verbose
)
}
\description{
Extract Parameter Distribution Densities
}
\keyword{internal}
