% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_checks.R
\name{.extract_prior_specs_standalone}
\alias{.extract_prior_specs_standalone}
\title{Extract Prior Specifications (Standalone Version)}
\usage{
.extract_prior_specs_standalone(prior, coef_names, verbose = TRUE)
}
\description{
Extract Prior Specifications (Standalone Version)
}
\keyword{internal}
