% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesR2.R
\name{.extract_random_effects_variance_corrected}
\alias{.extract_random_effects_variance_corrected}
\title{Extract random effects variance - CORRECTED}
\usage{
.extract_random_effects_variance_corrected(object, verbose)
}
\description{
Extract random effects variance - CORRECTED
}
\keyword{internal}
