% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_checks.R
\name{.extract_response_densities}
\alias{.extract_response_densities}
\title{Extract Response Distribution Densities}
\usage{
.extract_response_densities(
  object,
  show_prior,
  show_data,
  ndraws,
  prior_ndraws,
  verbose
)
}
\description{
Extract Response Distribution Densities
}
\keyword{internal}
