% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed_binomial_enhanced.R
\name{.fit_aggressive_strategy}
\alias{.fit_aggressive_strategy}
\title{Aggressive strategy}
\usage{
.fit_aggressive_strategy(formula, data, verbose, ...)
}
\description{
Aggressive strategy
}
\keyword{internal}
