% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed_binomial_enhanced.R
\name{.fit_enhanced_strategy}
\alias{.fit_enhanced_strategy}
\title{Enhanced strategy}
\usage{
.fit_enhanced_strategy(formula, data, regularisation_strength, verbose, ...)
}
\description{
Enhanced strategy
}
\keyword{internal}
