% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed_binomial_enhanced.R
\name{.fit_minimal_strategy}
\alias{.fit_minimal_strategy}
\title{Minimal strategy}
\usage{
.fit_minimal_strategy(formula, data, verbose, ...)
}
\description{
Minimal strategy
}
\keyword{internal}
