% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed_binomial_enhanced.R
\name{.fit_with_inla_enhanced}
\alias{.fit_with_inla_enhanced}
\title{Core INLA fitting function}
\usage{
.fit_with_inla_enhanced(
  formula,
  data,
  intercept_prior,
  slope_prior,
  control_strategy = "standard",
  verbose = TRUE,
  ...
)
}
\description{
Core INLA fitting function
}
\keyword{internal}
