% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesR2.R
\name{.generate_posterior_epred_corrected}
\alias{.generate_posterior_epred_corrected}
\title{Generate posterior expected predicted values - CORRECTED for mixed models}
\usage{
.generate_posterior_epred_corrected(
  object,
  newdata,
  ndraws,
  is_mixed,
  is_original_data,
  verbose
)
}
\description{
Generate posterior expected predicted values - CORRECTED for mixed models
}
\keyword{internal}
