% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_checks.R
\name{.generate_prior_only_samples}
\alias{.generate_prior_only_samples}
\title{Generate Prior-Only Samples}
\usage{
.generate_prior_only_samples(X, prior_specs, family, ndraws, verbose = TRUE)
}
\description{
Generate Prior-Only Samples
}
\keyword{internal}
