% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_visualisation.R
\name{.generate_prior_predictive_samples}
\alias{.generate_prior_predictive_samples}
\title{Generate Prior Predictive Samples}
\usage{
.generate_prior_predictive_samples(formula, data, family, prior, n_samples)
}
\description{
Generate Prior Predictive Samples
}
\keyword{internal}
