% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesR2.R
\name{.generate_random_effects_corrected}
\alias{.generate_random_effects_corrected}
\title{Generate random effects for one posterior draw - CORRECTED}
\usage{
.generate_random_effects_corrected(object, re_setup, draw_index, verbose)
}
\description{
Generate random effects for one posterior draw - CORRECTED
}
\keyword{internal}
