% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_checks.R
\name{.generate_response_from_family}
\alias{.generate_response_from_family}
\title{Generate Response from Family Distribution}
\usage{
.generate_response_from_family(linear_pred, family_name, n_obs)
}
\description{
Generate Response from Family Distribution
}
\keyword{internal}
