% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin_model_lab.R
\name{.infer_recommended_families}
\alias{.infer_recommended_families}
\title{Infer recommended families from a response vector}
\usage{
.infer_recommended_families(y, tol = 1e-08)
}
\description{
Infer recommended families from a response vector
}
\keyword{internal}
