% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_builder.R
\name{.mb_characterise_response}
\alias{.mb_characterise_response}
\title{Characterise Response Variable}
\usage{
.mb_characterise_response(y, var_name)
}
\description{
Characterise Response Variable
}
\keyword{internal}
