% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_builder.R
\name{.mb_get_priors}
\alias{.mb_get_priors}
\title{Get Prior Specifications}
\usage{
.mb_get_priors(family_choice, predictors, random_effects)
}
\description{
Get Prior Specifications
}
\keyword{internal}
