% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_builder.R
\name{.mb_get_random_effects}
\alias{.mb_get_random_effects}
\title{Get Random Effects Structure}
\usage{
.mb_get_random_effects(data, predictors)
}
\description{
Get Random Effects Structure
}
\keyword{internal}
