% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_visualisation.R
\name{.parse_prior_specification}
\alias{.parse_prior_specification}
\title{Parse Prior Specification}
\usage{
.parse_prior_specification(prior, parameter)
}
\description{
Parse Prior Specification
}
\keyword{internal}
