% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesR2.R
\name{.prepare_random_effects_corrected}
\alias{.prepare_random_effects_corrected}
\title{Prepare random effects structure - CORRECTED}
\usage{
.prepare_random_effects_corrected(object, newdata, verbose)
}
\description{
Prepare random effects structure - CORRECTED
}
\keyword{internal}
