% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_diagnostics.R
\name{.print_diagnostic_summary}
\alias{.print_diagnostic_summary}
\title{Print Diagnostic Summary}
\usage{
.print_diagnostic_summary(overall_status, results)
}
\description{
Print Diagnostic Summary
}
\keyword{internal}
