% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_effects.R
\name{.qbrms__compute_geometric_covariance}
\alias{.qbrms__compute_geometric_covariance}
\title{Compute geometric covariance from design matrix properties}
\usage{
.qbrms__compute_geometric_covariance(object, V_current)
}
\description{
Compute geometric covariance from design matrix properties
}
\keyword{Compute}
\keyword{covariance}
\keyword{design}
\keyword{from}
\keyword{geometric}
\keyword{internal}
\keyword{matrix}
\keyword{properties}
