% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_effects.R
\name{.qbrms__compute_leverage_uncertainty}
\alias{.qbrms__compute_leverage_uncertainty}
\title{Compute leverage-aware confidence intervals}
\usage{
.qbrms__compute_leverage_uncertainty(Xgrid, V, object, prob = 0.95)
}
\description{
Compute leverage-aware confidence intervals
}
\keyword{internal}
