% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_effects.R
\name{.qbrms__compute_ols_covariance}
\alias{.qbrms__compute_ols_covariance}
\title{Compute OLS covariance matrix as fallback}
\usage{
.qbrms__compute_ols_covariance(object)
}
\description{
Compute OLS covariance matrix as fallback
}
\keyword{internal}
