% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_effects.R
\name{.qbrms__enhance_covariance_matrix}
\alias{.qbrms__enhance_covariance_matrix}
\title{Enhance covariance matrix to ensure proper correlations}
\usage{
.qbrms__enhance_covariance_matrix(object, V_current, verbose = TRUE)
}
\description{
Enhance covariance matrix to ensure proper correlations
}
\keyword{internal}
