% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesR2.R
\name{.reconstruct_predictions_corrected}
\alias{.reconstruct_predictions_corrected}
\title{Reconstruct predictions manually - CORRECTED for mixed models}
\usage{
.reconstruct_predictions_corrected(object, newdata, ndraws, is_mixed, verbose)
}
\description{
Reconstruct predictions manually - CORRECTED for mixed models
}
\keyword{internal}
