% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed_binomial_enhanced.R
\name{.remove_random_effects}
\alias{.remove_random_effects}
\title{Remove random effects from formula}
\usage{
.remove_random_effects(formula)
}
\description{
Remove random effects from formula
}
\keyword{internal}
