% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesR2.R
\name{.sample_from_inla_random_effects}
\alias{.sample_from_inla_random_effects}
\title{Sample from INLA random effects posteriors - CORRECTED}
\usage{
.sample_from_inla_random_effects(object, re_setup, verbose)
}
\description{
Sample from INLA random effects posteriors - CORRECTED
}
\keyword{internal}
