% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_visualisation.R
\name{.sample_from_prior}
\alias{.sample_from_prior}
\title{Sample from Prior Distribution}
\usage{
.sample_from_prior(prior_info, n_samples)
}
\description{
Sample from Prior Distribution
}
\keyword{internal}
