% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesR2.R
\name{.summarise_bayes_r2}
\alias{.summarise_bayes_r2}
\title{Summarise Bayesian R-squared values}
\usage{
.summarise_bayes_r2(r2_values, robust, probs)
}
\description{
Summarise Bayesian R-squared values
}
\keyword{internal}
