% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed_binomial_enhanced.R
\name{.validate_qbrmb_inputs}
\alias{.validate_qbrmb_inputs}
\title{Validate qbrmb inputs}
\usage{
.validate_qbrmb_inputs(formula, data, family, verbose)
}
\description{
Validate qbrmb inputs
}
\keyword{internal}
