% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{drop_random_effects}
\alias{drop_random_effects}
\title{Drop Random Effects from Formula}
\usage{
drop_random_effects(formula)
}
\arguments{
\item{formula}{A model formula that may contain random effects.}
}
\value{
Formula with random effects terms removed.
}
\description{
Remove random effects terms from a model formula.
}
