% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{exponential}
\alias{exponential}
\alias{prior_exponential}
\title{Exponential Distribution (Prior or Family)}
\usage{
exponential(rate_or_link = "log", link = NULL, ...)

prior_exponential(rate_or_link = 1)
}
\arguments{
\item{rate_or_link}{Rate parameter (numeric) or link function (character).}

\item{link}{Optional link function (if acting as family).}

\item{...}{Additional arguments.}
}
\value{
A family object or prior object depending on inputs.
}
\description{
Exponential Distribution (Prior or Family)
}
