% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_tables.R
\name{extract_model_info}
\alias{extract_model_info}
\title{Extract Model Information for HTML Table}
\usage{
extract_model_info(
  model,
  ci.lvl,
  rope,
  show.rope,
  show.p_sig,
  show.pd,
  show.bf,
  verbose = FALSE
)
}
\description{
Extract Model Information for HTML Table
}
\keyword{internal}
