% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_model_metrics}
\alias{extract_model_metrics}
\title{Extract Model Metrics}
\usage{
extract_model_metrics(inla_fit)
}
\arguments{
\item{inla_fit}{INLA model object.}
}
\value{
A list of metrics.
}
\description{
Extract DIC, WAIC and other metrics from an INLA fit.
}
\keyword{internal}
