% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/families.R
\name{extract_routing_info}
\alias{extract_routing_info}
\title{Extract Routing Information from Family Specification}
\usage{
extract_routing_info(family_spec)
}
\arguments{
\item{family_spec}{Family specification with routing information}
}
\value{
List containing routing details
}
\description{
Extract Routing Information from Family Specification
}
\keyword{internal}
