% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional-families.R
\name{family_info}
\alias{family_info}
\title{Get Family Documentation}
\usage{
family_info(family_name)
}
\arguments{
\item{family_name}{Name of the family}
}
\value{
Character string with family information
}
\description{
Returns documentation and usage information for a family.
}
