% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fitted.qbrms_fit}
\alias{fitted.qbrms_fit}
\title{Extract fitted values from qbrms models}
\usage{
\method{fitted}{qbrms_fit}(object, ...)
}
\arguments{
\item{object}{A qbrms_fit object}

\item{...}{Additional arguments (currently unused)}
}
\value{
Numeric vector of fitted values
}
\description{
Extract fitted values from qbrms models
}
\examples{
\dontrun{
fit <- qbrms(mpg ~ hp, data = mtcars, family = gaussian())
fitted_values <- fitted(fit)
}

}
