% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinal_tmb.R
\name{fitted.tmb_ordinal_qbrms_fit}
\alias{fitted.tmb_ordinal_qbrms_fit}
\title{Fitted Values Method for TMB Ordinal Fits}
\usage{
\method{fitted}{tmb_ordinal_qbrms_fit}(object, ...)
}
\arguments{
\item{object}{A tmb_ordinal_qbrms_fit object}

\item{...}{Additional arguments}
}
\value{
Fitted values
}
\description{
Fitted Values Method for TMB Ordinal Fits
}
