% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_methods.R
\name{format_digits}
\alias{format_digits}
\title{Format numerical values to specified digits}
\usage{
format_digits(x, digits = 2)
}
\description{
Format numerical values to specified digits
}
\keyword{internal}
