% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_methods.R
\name{format_numeric_df}
\alias{format_numeric_df}
\title{Format data frame with numerical columns}
\usage{
format_numeric_df(df, digits = 2)
}
\description{
Format data frame with numerical columns
}
\keyword{internal}
