% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{gamma}
\alias{gamma}
\alias{gamma_prior}
\title{Gamma Distribution (Prior or Family)}
\usage{
gamma(shape_or_link = "log", rate = 1, link = NULL, ...)

gamma_prior(shape_or_link = 2, rate = 1)
}
\arguments{
\item{shape_or_link}{Shape parameter (numeric) or link function (character).}

\item{rate}{Rate parameter.}

\item{link}{Optional link function (if acting as family).}

\item{...}{Additional arguments.}
}
\value{
A family object or prior object depending on inputs.
}
\description{
Gamma Distribution (Prior or Family)
}
