% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional-families.R
\name{Gamma_family}
\alias{Gamma_family}
\alias{Gamma}
\title{Gamma family (GLM-style)}
\usage{
Gamma(link = "log")
}
\arguments{
\item{link}{Link function (default: "log")}
}
\value{
A family object
}
\description{
Gamma family constructor to avoid conflict with base::gamma.
}
