% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{gamma_prior}
\alias{gamma_prior}
\title{Specify Gamma Prior Distribution}
\usage{
gamma_prior(shape = 2, rate = 1)
}
\arguments{
\item{shape}{Shape parameter (must be positive)}

\item{rate}{Rate parameter (must be positive)}
}
\value{
A prior distribution object
}
\description{
Specify Gamma Prior Distribution
}
