% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_methods.R
\name{generate_posterior_predictive_samples}
\alias{generate_posterior_predictive_samples}
\title{Generate posterior predictive samples}
\usage{
generate_posterior_predictive_samples(object, ndraws = 100)
}
\description{
Generate posterior predictive samples
}
\keyword{internal}
