% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_tables.R
\name{generate_table_css}
\alias{generate_table_css}
\title{Generate CSS for different table styles}
\usage{
generate_table_css(style, font.family, font.size, header.bg, stripe.bg)
}
\description{
Generate CSS for different table styles
}
\keyword{internal}
