% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{get_default_prior}
\alias{get_default_prior}
\title{Get Default Prior for Parameter Class}
\usage{
get_default_prior(class)
}
\arguments{
\item{class}{Parameter class}
}
\value{
A default prior for that class
}
\description{
Get Default Prior for Parameter Class
}
