% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_parsing.R
\name{get_predictor_variables}
\alias{get_predictor_variables}
\title{Get Predictor Variables from Formula}
\usage{
get_predictor_variables(formula, data)
}
\arguments{
\item{formula}{Model formula}

\item{data}{Data frame}
}
\value{
List with predictor variable information
}
\description{
Extract predictor variables from formula, categorised by type
}
\keyword{internal}
