% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinal_tmb.R
\name{is_ordinal}
\alias{is_ordinal}
\title{Check if family is ordinal}
\usage{
is_ordinal(family)
}
\arguments{
\item{family}{Family specification}
}
\value{
Logical indicating if family is ordinal
}
\description{
Check if family is ordinal
}
\keyword{internal}
