% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{lognormal_prior}
\alias{lognormal_prior}
\title{Specify Lognormal Prior Distribution}
\usage{
lognormal_prior(meanlog = 0, sdlog = 1)
}
\arguments{
\item{meanlog}{Mean of log scale}

\item{sdlog}{Standard deviation of log scale (must be positive)}
}
\value{
A prior distribution object
}
\description{
Specify Lognormal Prior Distribution
}
