% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting.R
\name{model_fitting}
\alias{model_fitting}
\title{Model Fitting Functions for qbrms Package}
\description{
Core model fitting functionality with qbrm alias and multinomial support
}
\details{
This file contains the main qbrms/qbrm function and all supporting functions
}
\keyword{internal}
