% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting.R
\name{neg_binomial}
\alias{neg_binomial}
\title{Negative Binomial Family}
\usage{
neg_binomial()
}
\value{
An object of class \code{"family"} representing the Negative Binomial distribution.
}
\description{
Negative Binomial Family
}
