% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting.R
\name{negbinomial}
\alias{negbinomial}
\title{Negative Binomial Family (Alias)}
\usage{
negbinomial()
}
\value{
An object of class \code{"family"} representing the Negative Binomial distribution.
}
\description{
Negative Binomial Family (Alias)
}
