% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{normal}
\alias{normal}
\title{Specify Normal Prior Distribution}
\usage{
normal(mean = 0, sd = 1)
}
\arguments{
\item{mean}{Mean of the normal distribution (default 0)}

\item{sd}{Standard deviation of the normal distribution (default 1)}
}
\value{
A prior distribution object
}
\description{
Specify Normal Prior Distribution
}
