% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{null-coalesce}
\alias{null-coalesce}
\alias{\%||\%}
\title{Null Coalescing Operator}
\usage{
x \%||\% y
}
\arguments{
\item{x}{First value.}

\item{y}{Second value.}
}
\value{
x if not NULL, otherwise y.
}
\description{
Returns the first non-NULL value.
}
\keyword{internal}
