% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_parsing.R
\name{parse_brms_formula}
\alias{parse_brms_formula}
\title{Parse brms Formula Objects}
\usage{
parse_brms_formula(formula)
}
\arguments{
\item{formula}{Formula or brms formula object}
}
\value{
List with parsed formula information
}
\description{
Parse brms formula objects including bf() specifications
}
\keyword{internal}
