% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_effects.R
\name{plot.brms_conditional_effects}
\alias{plot.brms_conditional_effects}
\title{Plot Conditional Effects}
\usage{
\method{plot}{brms_conditional_effects}(
  x,
  ask = TRUE,
  spaghetti_args = list(),
  line_args = list(),
  ribbon = TRUE,
  ribbon_args = list(),
  ...
)
}
\arguments{
\item{x}{Object of class "brms_conditional_effects"}

\item{ask}{Logical, whether to ask before showing each plot}

\item{spaghetti_args}{List of arguments for spaghetti plot appearance}

\item{line_args}{List of arguments for main line appearance}

\item{ribbon}{Logical, whether to show confidence ribbon}

\item{ribbon_args}{List of arguments for ribbon appearance}

\item{...}{Additional arguments passed to ggplot2}
}
\value{
List of ggplot objects (returned invisibly)
}
\description{
Plot method for conditional effects from qbrms models.
}
