% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_diagnostics.R
\name{plot.qbrms_diagnostics}
\alias{plot.qbrms_diagnostics}
\title{Plot Method for Diagnostics}
\usage{
\method{plot}{qbrms_diagnostics}(x, which = "all", ...)
}
\arguments{
\item{x}{A qbrms_diagnostics object}

\item{which}{Character vector specifying which plots to show}

\item{...}{Additional arguments (unused)}
}
\description{
Plot Method for Diagnostics
}
