% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_analysis_tools.R
\name{plot.qbrms_p_significance}
\alias{plot.qbrms_p_significance}
\title{Plot Method for Enhanced p_significance}
\usage{
\method{plot}{qbrms_p_significance}(x, ...)

\method{plot}{qbrms_p_significance}(x, ...)
}
\arguments{
\item{x}{A \code{qbrms_p_significance} object from \code{p_significance()}.}

\item{...}{Additional arguments passed to ggplot2 functions.}
}
\value{
A ggplot2 object.

A ggplot2 object.
}
\description{
Create a visual plot of probability of practical significance results.

Create a visual plot of probability of practical significance results.
}
