% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting.R
\name{poisson_trick_multinomial}
\alias{poisson_trick_multinomial}
\title{Poisson Trick for Multinomial}
\usage{
poisson_trick_multinomial()
}
\value{
An internal family object for multinomial estimation via Poisson.
}
\description{
Poisson Trick for Multinomial
}
\keyword{internal}
