% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_methods.R
\name{pp_check.qbrms_fit}
\alias{pp_check.qbrms_fit}
\title{Posterior predictive checks for qbrms_fit objects}
\usage{
\method{pp_check}{qbrms_fit}(object, type = "dens_overlay", ndraws = 100, seed = NULL, ...)
}
\arguments{
\item{object}{A qbrms_fit object}

\item{type}{Type of plot ("dens_overlay", "hist", "scatter", "scatter_avg")}

\item{ndraws}{Number of posterior draws to use}

\item{seed}{Random seed for reproducibility}

\item{...}{Additional arguments}
}
\description{
Posterior predictive checks for qbrms_fit objects
}
