% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_diagnostics.R
\name{print.qbrms_diagnostics}
\alias{print.qbrms_diagnostics}
\title{Print Method for Diagnostics}
\usage{
\method{print}{qbrms_diagnostics}(x, ...)
}
\arguments{
\item{x}{A qbrms_diagnostics object}

\item{...}{Additional arguments (unused)}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Print Method for Diagnostics
}
