% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfold_cv.R
\name{print.qbrms_kfold}
\alias{print.qbrms_kfold}
\title{Print Method for qbrms_kfold Objects}
\usage{
\method{print}{qbrms_kfold}(x, ...)
}
\arguments{
\item{x}{A qbrms_kfold object}

\item{...}{Additional arguments (unused)}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Print Method for qbrms_kfold Objects
}
