% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo_compare.R
\name{print.qbrms_loo_compare}
\alias{print.qbrms_loo_compare}
\title{Print Method for qbrms_loo_compare Objects}
\usage{
\method{print}{qbrms_loo_compare}(x, ...)
}
\arguments{
\item{x}{A qbrms_loo_compare object}

\item{...}{Additional arguments (unused)}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Print Method for qbrms_loo_compare Objects
}
