% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_builder.R
\name{print.qbrms_model_spec}
\alias{print.qbrms_model_spec}
\title{Print Method for qbrms_model_spec}
\usage{
\method{print}{qbrms_model_spec}(x, ...)
}
\arguments{
\item{x}{A qbrms_model_spec object}

\item{...}{Additional arguments (unused)}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Print Method for qbrms_model_spec
}
