% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_elicitation.R
\name{print.qbrms_prior_code}
\alias{print.qbrms_prior_code}
\title{Print method for qbrms_prior_code objects}
\usage{
\method{print}{qbrms_prior_code}(x, ...)
}
\arguments{
\item{x}{A qbrms_prior_code object}

\item{...}{Additional arguments passed to cat}
}
\description{
Print method for qbrms_prior_code objects
}
