% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{print.qbrms_prior_dist}
\alias{print.qbrms_prior_dist}
\title{Print Prior Distribution Objects}
\usage{
\method{print}{qbrms_prior_dist}(x, ...)
}
\arguments{
\item{x}{A qbrms_prior_dist object}

\item{...}{Unused}
}
\description{
Print Prior Distribution Objects
}
