% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{print.qbrms_prior_list}
\alias{print.qbrms_prior_list}
\title{Print Prior List Objects}
\usage{
\method{print}{qbrms_prior_list}(x, ...)
}
\arguments{
\item{x}{A qbrms_prior_list object}

\item{...}{Unused}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print Prior List Objects
}
