% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{print.qbrms_prior_spec}
\alias{print.qbrms_prior_spec}
\title{Print Prior Specification Objects}
\usage{
\method{print}{qbrms_prior_spec}(x, ...)
}
\arguments{
\item{x}{A qbrms_prior_spec object}

\item{...}{Unused}
}
\description{
Print Prior Specification Objects
}
