% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributional_syntax.R
\name{print.qbrmsformula}
\alias{print.qbrmsformula}
\title{Print method for qbrms formulas}
\usage{
\method{print}{qbrmsformula}(x, ...)
}
\arguments{
\item{x}{A qbrmsformula object}

\item{...}{Unused}
}
\description{
Print method for qbrms formulas
}
