% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{print.summary.qbrms_fit}
\alias{print.summary.qbrms_fit}
\title{Print Method for summary.qbrms_fit Objects}
\usage{
\method{print}{summary.qbrms_fit}(x, ...)
}
\arguments{
\item{x}{A \code{summary.qbrms_fit} object.}

\item{...}{Additional arguments (currently unused).}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Print Method for summary.qbrms_fit Objects
}
