% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinal_tmb.R
\name{print.tmb_ordinal_qbrms_fit}
\alias{print.tmb_ordinal_qbrms_fit}
\title{Print Method for TMB Ordinal Fits}
\usage{
\method{print}{tmb_ordinal_qbrms_fit}(x, digits = 2, ...)
}
\arguments{
\item{x}{A tmb_ordinal_qbrms_fit object}

\item{digits}{Number of decimal places for output}

\item{...}{Additional arguments}
}
\value{
Invisibly returns the object
}
\description{
Print Method for TMB Ordinal Fits
}
