% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_elicitation.R
\name{prior_code}
\alias{prior_code}
\title{Format priors as qbrms prior() code}
\usage{
prior_code(build, object_name = "priors", digits = 3, include_comments = TRUE)
}
\arguments{
\item{build}{A 'qbrms_prior_build' returned by prior_build_from_beliefs()}

\item{object_name}{Name of the object on the left-hand side (default "priors")}

\item{digits}{Number of decimal places to print}

\item{include_comments}{Logical; if TRUE, prepend a short comment header}
}
\value{
A single character string containing formatted R code
}
\description{
Format priors as qbrms prior() code
}
