% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{prior_exponential}
\alias{prior_exponential}
\title{Specify Exponential Prior Distribution}
\usage{
prior_exponential(rate = 1)
}
\arguments{
\item{rate}{Rate parameter for the exponential distribution (default 1)}
}
\value{
A prior distribution object
}
\description{
Specify Exponential Prior Distribution
}
