% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbosity.R
\name{qbrms_fit_log}
\alias{qbrms_fit_log}
\title{Get captured fit log from a qbrms object (if available)}
\usage{
qbrms_fit_log(x)
}
\arguments{
\item{x}{A qbrms_fit / qbrmO_fit object returned by qbrm()/qbrms()}
}
\value{
A character vector of captured console lines, or NULL if none
}
\description{
Get captured fit log from a qbrms object (if available)
}
