% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/families.R
\name{requires_routing}
\alias{requires_routing}
\title{Check if Family Requires Routing to Specialist Implementation}
\usage{
requires_routing(family_spec)
}
\arguments{
\item{family_spec}{Family specification from convert_family_to_inla()}
}
\value{
Logical indicating if routing is required
}
\description{
Check if Family Requires Routing to Specialist Implementation
}
\keyword{internal}
