% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{residuals.qbrms_fit}
\alias{residuals.qbrms_fit}
\title{Extract residuals from qbrms models}
\usage{
\method{residuals}{qbrms_fit}(object, type = "response", ...)
}
\arguments{
\item{object}{A qbrms_fit object}

\item{type}{Character string indicating type of residuals (default: "response")}

\item{...}{Additional arguments (currently unused)}
}
\value{
Numeric vector of residuals
}
\description{
Extract residuals from qbrms models
}
\examples{
\dontrun{
fit <- qbrms(mpg ~ hp, data = mtcars, family = gaussian())
resid_values <- residuals(fit, type = "response")
}

}
