% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_methods.R
\name{set_qbrms_digits}
\alias{set_qbrms_digits}
\title{Set default formatting options for qbrms output}
\usage{
set_qbrms_digits(digits = 2)
}
\arguments{
\item{digits}{Default number of decimal places (default 2)}
}
\description{
Set default formatting options for qbrms output
}
\examples{
\dontrun{
# Set default to 3 decimal places
set_qbrms_digits(3)

# Reset to default
set_qbrms_digits(2)
}
}
