% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinal_tmb.R
\name{setup_ordinal_tmb}
\alias{setup_ordinal_tmb}
\title{Set Up TMB Model Object}
\usage{
setup_ordinal_tmb(tmb_data, prior_params, control, verbose = TRUE)
}
\arguments{
\item{tmb_data}{Prepared TMB data structure}

\item{prior_params}{Processed prior parameters}

\item{control}{Control parameters for TMB}

\item{verbose}{Logical; print progress messages}
}
\value{
List with TMB setup information
}
\description{
Set Up TMB Model Object
}
\keyword{internal}
