% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting.R
\name{skew_normal}
\alias{skew_normal}
\title{Skew Normal Family}
\usage{
skew_normal()
}
\value{
An object of class \code{"family"} representing the Skew Normal distribution.
}
\description{
Skew Normal Family
}
