% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-test-utilities.R
\name{skip_if_not_installed}
\alias{skip_if_not_installed}
\title{Skip test if package not available}
\usage{
skip_if_not_installed(pkg)
}
\description{
Skip test if package not available
}
\keyword{internal}
