% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{student_t_prior}
\alias{student_t_prior}
\title{Specify Student's t Prior Distribution}
\usage{
student_t_prior(df = 3, location = 0, scale = 1)
}
\arguments{
\item{df}{Degrees of freedom (default 3)}

\item{location}{Location parameter (default 0)}

\item{scale}{Scale parameter (default 1)}
}
\value{
A prior distribution object
}
\description{
Specify Student's t Prior Distribution
}
