% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{uniform}
\alias{uniform}
\title{Specify Uniform Prior Distribution}
\usage{
uniform(min = -Inf, max = Inf)
}
\arguments{
\item{min}{Lower bound (default -Inf)}

\item{max}{Upper bound (default Inf)}
}
\value{
A prior distribution object
}
\description{
Specify Uniform Prior Distribution
}
