% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family_conversion.R, R/model_fitting.R
\name{vcov.inla}
\alias{vcov.inla}
\title{Variance-Covariance Matrix for INLA Objects}
\usage{
\method{vcov}{inla}(object, ...)

\method{vcov}{inla}(object, ...)
}
\arguments{
\item{object}{An INLA fit object.}

\item{...}{Additional arguments (currently unused).}
}
\value{
A variance-covariance matrix for the fixed effects.
}
\description{
Extract the variance-covariance matrix from an INLA fit object.
}
