% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_methods.R
\name{vcov.qbrms_fit}
\alias{vcov.qbrms_fit}
\title{Extract Variance-Covariance Matrix from qbrms Models}
\usage{
\method{vcov}{qbrms_fit}(object, ...)
}
\arguments{
\item{object}{A qbrms_fit object}

\item{...}{Additional arguments (unused)}
}
\value{
Variance-covariance matrix
}
\description{
Extract Variance-Covariance Matrix from qbrms Models
}
