% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinal_tmb.R
\name{vcov.tmb_ordinal_qbrms_fit}
\alias{vcov.tmb_ordinal_qbrms_fit}
\title{Variance-Covariance Matrix Method for TMB Ordinal Fits}
\usage{
\method{vcov}{tmb_ordinal_qbrms_fit}(object, ...)
}
\arguments{
\item{object}{A tmb_ordinal_qbrms_fit object}

\item{...}{Additional arguments}
}
\value{
Variance-covariance matrix
}
\description{
Variance-Covariance Matrix Method for TMB Ordinal Fits
}
