% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_tables.R
\name{view_table}
\alias{view_table}
\title{Display HTML Table in Viewer}
\usage{
view_table(x)
}
\arguments{
\item{x}{A \code{qbrms_html_table} created by \code{tab_model()}.}
}
\description{
Display HTML Table in Viewer
}
