% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional-families.R, R/families.R
\name{Beta}
\alias{Beta}
\title{Beta Family Constructor (Capital B)}
\usage{
Beta(link = "logit", link.phi = "log")

Beta(link = "logit", link.phi = "log")
}
\arguments{
\item{link}{Link function for the mean parameter (default: "logit")}

\item{link.phi}{Link function for precision parameter (default: "log")}
}
\value{
A family object of class \code{"family"}.

A family object for use with qbrms()
}
\description{
Beta distribution family for response variables in (0,1)
}
\examples{
\dontrun{
# Beta regression for proportions
fit <- qbrms(proportion ~ predictor, data = data, family = Beta())
}

}
