% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{beta_prior}
\alias{beta_prior}
\title{Specify Beta Prior Distribution}
\usage{
beta_prior(alpha = 1, beta = 1)
}
\arguments{
\item{alpha}{First shape parameter (must be positive)}

\item{beta}{Second shape parameter (must be positive)}
}
\value{
A prior distribution object
}
\description{
Specify Beta Prior Distribution
}
