% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{c.qbrms_prior_spec}
\alias{c.qbrms_prior_spec}
\title{Combine Multiple Prior Specifications}
\usage{
\method{c}{qbrms_prior_spec}(...)
}
\arguments{
\item{...}{Prior specification objects created by prior()}
}
\value{
A combined prior object
}
\description{
Combine Multiple Prior Specifications
}
\examples{
\dontrun{
# Multiple priors
priors <- c(
  prior(normal(0, 2.5), class = "Intercept"),
  prior(normal(0, 1), class = "b"),
  prior(cauchy(0, 1), class = "sd")
)
}
}
