% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_effects.R
\name{conditional_effects}
\alias{conditional_effects}
\title{Conditional effects (generic)}
\usage{
conditional_effects(object, ...)
}
\arguments{
\item{object}{A model object.}

\item{...}{Passed to methods.}
}
\value{
An object of class \code{"qbrms_conditional_effects"} containing
conditional effect estimates. The structure is method-dependent.
}
\description{
Compute one-dimensional conditional effects / marginal fitted values as a
predictor varies while other covariates are held fixed (typically at means /
modes). Methods should return an object that \code{plot()} can visualise.
}
