% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_effects.R
\name{conditional_effects}
\alias{conditional_effects}
\title{Conditional effects (generic)}
\usage{
conditional_effects(object, ...)
}
\arguments{
\item{object}{A model object.}

\item{...}{Passed to methods.}
}
\description{
Generic for computing one-dimensional conditional effects / marginal fitted
values as a predictor varies (others held fixed). Methods should return a
\pkg{ggplot2} object (or a named list of them) that users can further modify.
}
