% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family_conversion.R
\name{create_quantile_fit}
\alias{create_quantile_fit}
\title{Create A Quantile Regression Fitted Object for 'INLA'}
\usage{
create_quantile_fit(formula, data, quantile = 0.5, verbose = TRUE)
}
\description{
Create A Quantile Regression Fitted Object for 'INLA'
}
\keyword{internal}
