% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional-families.R
\name{exponential}
\alias{exponential}
\title{Exponential Survival Family}
\usage{
exponential(link = "log")
}
\arguments{
\item{link}{Link function for rate parameter (default: "log")}
}
\value{
A family object for use with qbrms()
}
\description{
Exponential distribution for survival analysis (special case of Weibull with shape=1).
}
\examples{
\dontrun{
fit <- qbrms(time ~ treatment, data = survival_data, family = exponential())
fit2 <- qbrms(time ~ treatment, data = survival_data, family = exponential(link = "log"))
}

}
