% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional-families.R
\name{gen_student_t}
\alias{gen_student_t}
\title{Generalized t Family}
\usage{
gen_student_t(link = "identity", link.sigma = "log", link.nu = "log")
}
\arguments{
\item{link}{Link function for location}

\item{link.sigma}{Link function for scale}

\item{link.nu}{Link function for degrees of freedom}
}
\value{
A family object
}
\description{
Generalized t Family
}
