% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional-families.R
\name{gev}
\alias{gev}
\alias{gumbel}
\title{Generalized Extreme Value Family}
\usage{
gev(link = "identity", link.sigma = "log", link.xi = "identity")

gumbel(link = "identity", link.sigma = "log")
}
\arguments{
\item{link}{Link function for location}

\item{link.sigma}{Link function for scale}

\item{link.xi}{Link function for shape}
}
\value{
A family object of class \code{"family"}.
}
\description{
Generalized Extreme Value Family
}
