% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional-families.R
\name{hurdle_families}
\alias{hurdle_families}
\alias{hurdle_poisson}
\alias{hurdle_negbinomial}
\title{Hurdle Families for Two-Part Models}
\usage{
hurdle_poisson(link = "log", link.hu = "logit")

hurdle_negbinomial(link = "log", link.hu = "logit")
}
\arguments{
\item{link}{Link function for count component}

\item{link.hu}{Link function for hurdle component}
}
\value{
A family object of class \code{"family"}.
}
\description{
Hurdle Families for Two-Part Models
}
