% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional-families.R
\name{iid}
\alias{iid}
\title{IID Random Effects}
\usage{
iid(scale.model = "log", diagonal = 1e-06)
}
\arguments{
\item{scale.model}{Scaling model for precision}

\item{diagonal}{Diagonal precision matrix structure}
}
\value{
A family object
}
\description{
IID Random Effects
}
