% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional-families.R
\name{laplace}
\alias{laplace}
\alias{double_exponential}
\title{Laplace (Double Exponential) Family}
\usage{
laplace(link = "identity", link.sigma = "log")

double_exponential(link = "identity", link.sigma = "log")
}
\arguments{
\item{link}{Link function for location}

\item{link.sigma}{Link function for scale}
}
\value{
A family object
}
\description{
Laplace (Double Exponential) Family
}
