% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/families.R, R/priors.R
\name{lognormal}
\alias{lognormal}
\alias{lognormal_prior}
\title{Lognormal Family Constructor}
\usage{
lognormal(meanlog_or_link = "identity", sdlog = 1, link = NULL, ...)

lognormal(meanlog_or_link = "identity", sdlog = 1, link = NULL, ...)

lognormal_prior(meanlog_or_link = 0, sdlog = 1)
}
\arguments{
\item{meanlog_or_link}{Mean on log scale (numeric) or link function (character).}

\item{sdlog}{SD on log scale (numeric).}

\item{link}{Optional link function (if acting as family).}

\item{...}{Additional arguments.}
}
\value{
A family object for use with qbrms()

A family object or prior object depending on inputs.
}
\description{
Lognormal distribution family for positive continuous responses
}
\examples{
\dontrun{
# Lognormal regression 
fit <- qbrms(response ~ predictor, data = data, family = lognormal())
}

}
