% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_methods.R
\name{pp_check}
\alias{pp_check}
\title{Posterior and Prior Predictive Checks (generic)}
\usage{
pp_check(object, ...)
}
\arguments{
\item{object}{A model object}

\item{...}{Additional arguments passed to methods}
}
\description{
Posterior and Prior Predictive Checks (generic)
}
