% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{prior}
\alias{prior}
\title{Specify Prior for Model Parameters}
\usage{
prior(prior, class = "b", coef = NULL, group = NULL)
}
\arguments{
\item{prior}{A prior distribution object.}

\item{class}{Parameter class ("Intercept", "b", "sd", etc.)}

\item{coef}{Specific coefficient name (optional)}

\item{group}{Specific group name for random effects (optional)}
}
\value{
A prior specification object
}
\description{
Specify Prior for Model Parameters
}
