% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional-families.R
\name{random_walk_families}
\alias{random_walk_families}
\alias{rw1}
\alias{rw2}
\title{Random Walk Families}
\usage{
rw1(scale.model = "log", diagonal = 1e-06)

rw2(scale.model = "log", diagonal = 1e-06)
}
\arguments{
\item{scale.model}{Scaling model for precision}

\item{diagonal}{Diagonal precision matrix structure}
}
\value{
A family object of class \code{"family"}.
}
\description{
Random Walk Families
}
