% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional-families.R
\name{simplex}
\alias{simplex}
\title{Simplex Family for Compositional Data}
\usage{
simplex(link = "logit", link.precision = "log")
}
\arguments{
\item{link}{Link function for the mean (default: "logit")}

\item{link.precision}{Link function for precision parameter (default: "log")}
}
\value{
A family object of class \code{"family"}.
}
\description{
Simplex Family for Compositional Data
}
