% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed_binomial_enhanced.R
\name{summary.qbrmb_fit}
\alias{summary.qbrmb_fit}
\title{Summary method for qbrmb_fit objects}
\usage{
\method{summary}{qbrmb_fit}(object, digits = 2, show_diagnostics = NULL, ...)
}
\arguments{
\item{object}{A qbrmb_fit object}

\item{digits}{Number of digits to display}

\item{show_diagnostics}{Logical; show diagnostic information}

\item{...}{Additional arguments passed to print}
}
\description{
Summary method for qbrmb_fit objects
}
