% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_methods.R
\name{summary.qbrms_fit}
\alias{summary.qbrms_fit}
\title{Summary Method for qbrms Models}
\usage{
\method{summary}{qbrms_fit}(object, ..., digits = 2)
}
\arguments{
\item{object}{A qbrms_fit object}

\item{...}{Additional arguments}

\item{digits}{Number of digits for output (default 2)}
}
\description{
Summary Method for qbrms Models
}
