% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional-families.R
\name{weibull}
\alias{weibull}
\title{Weibull Survival Family}
\usage{
weibull(link = "log", link.shape = "log")
}
\arguments{
\item{link}{Link function for scale (default: "log")}

\item{link.shape}{Link function for shape (default: "log")}
}
\value{
A family object of class \code{"family"}.
}
\description{
Weibull Survival Family
}
