% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional-families.R
\name{zero_inflated_negbinomial}
\alias{zero_inflated_negbinomial}
\alias{zinb}
\title{Zero-Inflated Negative Binomial Family}
\usage{
zero_inflated_negbinomial(link = "log", link.zi = "logit")

zinb(link = "log", link.zi = "logit")
}
\arguments{
\item{link}{Link function for mean (default: "log")}

\item{link.zi}{Link function for zero-inflation (default: "logit")}
}
\value{
A family object
}
\description{
Zero-Inflated Negative Binomial Family
}
