% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional-families.R
\name{zero_inflated_poisson}
\alias{zero_inflated_poisson}
\alias{zip}
\title{Zero-Inflated Poisson Family}
\usage{
zero_inflated_poisson(link = "log", link.zi = "logit")

zip()
}
\arguments{
\item{link}{Link function for mean (default: "log")}

\item{link.zi}{Link function for zero-inflation (default: "logit")}
}
\value{
A family object
}
\description{
Zero-Inflated Poisson Family
}
